refactor(llama): extract LlamaGenerationDriver from LlamaBackend (#407)
